Crate cpp_common

Implementation detail for the cpp crate.

The purpose of this crate is only to allow sharing code between the cpp_build and the cpp_macros crates.

Modules

flags

kw

Structs

Capture

Class

Closure

ClosureSig

FILE_HASH

OUT_DIR

RustInvocation

Enums

Macro

Constants

LIB_NAME

MSVC_LIB_NAME

STRUCT_METADATA_MAGIC
This constant is expected to be a unique string within the compiled binary which precedes a definition of the metadata. It begins with rustcpp~metadata, which is printable to make it easier to locate when looking at a binary dump of the metadata.

VERSION